docs: expand WASM ABI v1 shapes and harden helpers
- Document v1 JSON request/response shapes for completions, code actions, navigation, and inlay hints\n- Harden nova-ext-abi guest helpers: avoid overflow-to-`usize::MAX` and return (0,0) on oversized buffers\n- Treat alloc-returned pointers as u32 bits in nova-ext wasm host to avoid sign-extension issues

+++ crates/nova-ext/src/wasm/runtime.rs
@@ -403,7 +403,7 @@ impl WasmPlugin {
         let req_len_i32 = i32::try_from(req_bytes.len()).unwrap_or(i32::MAX);
         let req_ptr = alloc
             .call(&mut store, req_len_i32)
-            .map_err(classify_call_error)? as usize;
+            .map_err(classify_call_error)? as u32 as usize;
 
         memory
             .write(&mut store, req_ptr, &req_bytes)
